requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
groq==0.13.1
spacy==3.8.3
jinja2==3.1.5
python-multipart==0.0.20
python-dotenv==1.0.1
